Time: 490 ms (69.37%), Space: 46.9 MB (69.37%) - LeetHub

+++ 493-reverse-pairs/493-reverse-pairs.cpp
@@ -4,7 +4,7 @@ public:
         int i=left, j=mid, res=0;
         while(i<=mid-1){
             long long val,k=2;
-            if(j<=right) val= k*(long long)arr[j];
+            if(j<=right) val= k*arr[j];
             if(j==right+1) res+=j-mid, i++;
             else if(arr[i]>val) j++;
